Many fixes are made. Functions files are cleaned. Case 2035 are created.
